Moved graphql files to folders respective there types Query, Mutation, Subscription
