+++ constants/graphql/mutations/addFriendQuery.ts
