+++ constants/graphql/mutations/signUpQuery.ts
